requirements.txt
mysqlclient
docxtpl
Pillow
django-cleanup